server/requirements.txt
Flask
flask-cors
textblob
requests
python-dotenv